Dovetail.Core

Building core libraries

data CoreBuild ctx

Constructors

CoreBuild

Fields

buildInputs :: HashMap Text (ByteString, ByteString)

env :: Env ctx

buildModules :: forall ctx. Typeable ctx => CoreBuild ctx -> Interpret ctx ()

Package collections

allPackages :: Typeable ctx => CoreBuild ctx

minimal :: Typeable ctx => CoreBuild ctx

A smaller package collection which is the transitive closure of the arrays, strings, integers, numbers and math packages.

Individual packages

arrays :: Typeable ctx => CoreBuild ctx

assert :: Typeable ctx => CoreBuild ctx

bifunctors :: Typeable ctx => CoreBuild ctx

catenableLists :: Typeable ctx => CoreBuild ctx

console :: Typeable ctx => CoreBuild ctx

_const :: Typeable ctx => CoreBuild ctx

contravariant :: Typeable ctx => CoreBuild ctx

control :: Typeable ctx => CoreBuild ctx

distributive :: Typeable ctx => CoreBuild ctx

effect :: Typeable ctx => CoreBuild ctx

_either :: Typeable ctx => CoreBuild ctx

enums :: Typeable ctx => CoreBuild ctx

exceptions :: Typeable ctx => CoreBuild ctx

exists :: Typeable ctx => CoreBuild ctx

foldableTraversable :: Typeable ctx => CoreBuild ctx

orders :: Typeable ctx => CoreBuild ctx

free :: Typeable ctx => CoreBuild ctx

functions :: Typeable ctx => CoreBuild ctx

functors :: Typeable ctx => CoreBuild ctx

gen :: Typeable ctx => CoreBuild ctx

graphs :: Typeable ctx => CoreBuild ctx

identity :: Typeable ctx => CoreBuild ctx

integers :: Typeable ctx => CoreBuild ctx

invariant :: Typeable ctx => CoreBuild ctx

lazy :: Typeable ctx => CoreBuild ctx

lcg :: Typeable ctx => CoreBuild ctx

lists :: Typeable ctx => CoreBuild ctx

math :: Typeable ctx => CoreBuild ctx

_maybe :: Typeable ctx => CoreBuild ctx

_newtype :: Typeable ctx => CoreBuild ctx

nonempty :: Typeable ctx => CoreBuild ctx

numbers :: Typeable ctx => CoreBuild ctx

orderedCollections :: Typeable ctx => CoreBuild ctx

typelevelPrelude :: Typeable ctx => CoreBuild ctx

parallel :: Typeable ctx => CoreBuild ctx

partial :: Typeable ctx => CoreBuild ctx

prelude :: Typeable ctx => CoreBuild ctx

profunctor :: Typeable ctx => CoreBuild ctx

psciSupport :: Typeable ctx => CoreBuild ctx

quickcheck :: Typeable ctx => CoreBuild ctx

random :: Typeable ctx => CoreBuild ctx

record :: Typeable ctx => CoreBuild ctx

refs :: Typeable ctx => CoreBuild ctx

safeCoerce :: Typeable ctx => CoreBuild ctx

semirings :: Typeable ctx => CoreBuild ctx

st :: Typeable ctx => CoreBuild ctx

strings :: Typeable ctx => CoreBuild ctx

tailrec :: Typeable ctx => CoreBuild ctx

transformers :: Typeable ctx => CoreBuild ctx

tuples :: Typeable ctx => CoreBuild ctx

typeEquality :: Typeable ctx => CoreBuild ctx

unfoldable :: Typeable ctx => CoreBuild ctx

unsafeCoerce :: Typeable ctx => CoreBuild ctx

validation :: Typeable ctx => CoreBuild ctx